Moved host powers to host-events.js
