+++ src/scss/tracks.scss
@@ -96,7 +96,7 @@
 	.option-icons {
 		position: absolute;
 		z-index: 108;
-		right: 95px;
+		right: 73px;
 		left: auto;
 		top: 10px;
 		background: #5b9cc4;
